Add deletion route and make feedback user-based.

+++ app/controllers/feedback_items_controller.rb
@@ -3,7 +3,7 @@ class FeedbackItemsController < ApplicationController
   def update_feedback_items
     begin
       batch_params["feedback_items"].each do |patched_item|
-        item = FeedbackItem.where(document_inference_id: patched_item["document_inference_id"]).first_or_create
+        item = FeedbackItem.where(document_inference_id: patched_item["document_inference_id"], user_id: patched_item["user_id"]).first_or_create
         item.assign_attributes(patched_item)
         item.save!
       end
@@ -15,13 +15,13 @@ class FeedbackItemsController < ApplicationController
 
   def delete_items
     batch_params["feedback_items"].each do |patched_item|
-      FeedbackItem.where(document_inference_id: patched_item["document_inference_id"]).destroy_all
+      FeedbackItem.where(document_inference_id: patched_item["document_inference_id"], user_id: patched_item["user_id"]).destroy_all
     end
   end
 
   private
 
   def batch_params
-    params.permit(feedback_items: [:id, :document_inference_id, :sentiment, :comment]).to_h
+    params.permit(feedback_items: [:id, :document_inference_id, :user_id, :sentiment, :comment]).to_h
   end
 end
